% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFunctions.R
\name{av_runShiny_addFunctions}
\alias{av_runShiny_addFunctions}
\title{Add experimental functions}
\usage{
av_runShiny_addFunctions()
}
\value{
Nothing
}
\description{
Adds functions in this package
}
\details{
Add new functions
}
