% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Correlations.R
\name{rolling_correlations}
\alias{rolling_correlations}
\title{Rolling correlation analysis}
\usage{
rolling_correlations(todo, rv)
}
\arguments{
\item{todo}{Data structure containing  command and assets}

\item{rv}{Copy of internal av_runShiny() state}
}
\value{
List of output items
}
\description{
Rolling correlation analysis
}
\details{
rolling_correlations
}
\seealso{
\code{\link[=av_runShiny_addFunctions]{av_runShiny_addFunctions()}}
}
